Refactor GenericEmbed to extract reusable embed code to avoid duplication
