+++ src/blocks/GenericEmbed/config.ts
@@ -6,6 +6,10 @@ export const GenericEmbedBlock: Block = {
   slug: 'genericEmbed',
   imageURL: '/thumbnail/GenericEmbedThumbnail.jpg',
   interfaceName: 'GenericEmbedBlock',
+  labels: {
+    singular: 'Generic Embed',
+    plural: 'Generic Embeds',
+  },
   fields: [
     {
       name: 'html',
@@ -14,7 +18,7 @@ export const GenericEmbedBlock: Block = {
       required: true,
       admin: {
         description:
-          'Helpful tip: <iframe> tags should have hardcoded height and width. You can use relative (100%) or pixel values (600px) for width. You must use pixel values for height.',
+          'For arbitrary HTML/iframe embeds. For videos use the Video Embed block, and for donation or form widgets (DonorBox, etc.) use the Form Embed block. Helpful tip: <iframe> tags should have hardcoded height and width. You can use relative (100%) or pixel values (600px) for width. You must use pixel values for height.',
       },
     },
     {
