+++ src/endpoints/seed/blocks/generic-embed.ts
@@ -3,12 +3,6 @@ import { RequiredDataFromCollectionSlug } from 'payload'
 export const genericEmbed: RequiredDataFromCollectionSlug<'pages'>['layout'] = [
   {
     blockType: 'genericEmbed',
-    html: '<iframe width="560" height="315" src="https://www.youtube.com/embed/dQw4w9WgXcQ" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" allowfullscreen></iframe>',
-    backgroundColor: 'white',
-    alignContent: 'center',
-  },
-  {
-    blockType: 'genericEmbed',
     html: '<blockquote class="twitter-tweet"><p lang="en" dir="ltr">Example embedded tweet content</p>&mdash; Twitter (@twitter) <a href="https://twitter.com/twitter/status/1">January 1, 2021</a></blockquote>',
     backgroundColor: 'brand-400',
     alignContent: 'left',
